pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "jwks-server"
version = "0.1.0"
description = "Basic JWKS server (FastAPI) with JWKS + /auth"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "fastapi>=0.110",
  "uvicorn[standard]>=0.27",
  "pyjwt[crypto]>=2.9",
  "cryptography>=43.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
  "pytest-cov>=5.0",
  "httpx>=0.27",
  "ruff>=0.5",
  "black>=24.3",
  "coverage-badge>=1.1.0",
]

[tool.pytest.ini_options]
addopts = "--maxfail=1 -q --cov=app --cov-report=term-missing --cov-report=xml"

[tool.black]
line-length = 100
target-version = ["py310"]

[tool.ruff]
line-length = 100
select = ["E", "F", "I"]